changed types to Uint8List

+++ android/src/main/java/com/zezo357/pytorch_lite/Pigeon.java
@@ -390,9 +390,9 @@ public class Pigeon {
     /**predicts abstract number input */
     void getPredictionCustom(@NonNull Long index, @NonNull List<Double> input, @NonNull List<Long> shape, @NonNull String dtype, @NonNull Result<List<Object>> result);
     /**predicts raw image but returns the raw net output */
-    void getRawImagePredictionList(@NonNull Long index, @NonNull double[] imageData, @NonNull Result<List<Double>> result);
+    void getRawImagePredictionList(@NonNull Long index, @NonNull byte[] imageData, @NonNull Result<List<Double>> result);
     /**predicts raw image but returns the raw net output */
-    void getRawImagePredictionListObjectDetection(@NonNull Long index, @NonNull double[] imageData, @NonNull Double minimumScore, @NonNull Double IOUThreshold, @NonNull Long boxesLimit, @NonNull Result<List<ResultObjectDetection>> result);
+    void getRawImagePredictionListObjectDetection(@NonNull Long index, @NonNull byte[] imageData, @NonNull Double minimumScore, @NonNull Double IOUThreshold, @NonNull Long boxesLimit, @NonNull Result<List<ResultObjectDetection>> result);
     /**predicts image but returns the raw net output */
     void getImagePredictionList(@NonNull Long index, @Nullable byte[] imageData, @Nullable List<byte[]> imageBytesList, @Nullable Long imageWidthForBytesList, @Nullable Long imageHeightForBytesList, @NonNull List<Double> mean, @NonNull List<Double> std, @NonNull Result<List<Double>> result);
     /**predicts image but returns the output detections */
@@ -477,7 +477,7 @@ public class Pigeon {
                 ArrayList<Object> wrapped = new ArrayList<Object>();
                 ArrayList<Object> args = (ArrayList<Object>) message;
                 Number indexArg = (Number) args.get(0);
-                double[] imageDataArg = (double[]) args.get(1);
+                byte[] imageDataArg = (byte[]) args.get(1);
                 Result<List<Double>> resultCallback =
                     new Result<List<Double>>() {
                       public void success(List<Double> result) {
@@ -508,7 +508,7 @@ public class Pigeon {
                 ArrayList<Object> wrapped = new ArrayList<Object>();
                 ArrayList<Object> args = (ArrayList<Object>) message;
                 Number indexArg = (Number) args.get(0);
-                double[] imageDataArg = (double[]) args.get(1);
+                byte[] imageDataArg = (byte[]) args.get(1);
                 Double minimumScoreArg = (Double) args.get(2);
                 Double IOUThresholdArg = (Double) args.get(3);
                 Number boxesLimitArg = (Number) args.get(4);
